pyproject.toml
[project]
name = "process-gpt-memento"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp==3.10.4",
    "aiosignal==1.3.1",
    "annotated-types==0.6.0",
    "anyio==4.2.0",
    "async-timeout==4.0.3",
    "attrs==23.2.0",
    "beautifulsoup4==4.12.3",
    "certifi==2024.2.2",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "dataclasses-json==0.6.4",
    "deprecated==1.2.14",
    "dirtyjson==1.0.8",
    "distro==1.9.0",
    "exceptiongroup==1.2.0",
    "fastapi==0.109.2",
    "filelock==3.13.1",
    "frozenlist==1.4.1",
    "fsspec==2024.2.0",
    "google-api-python-client==2.118.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.0",
    "greenlet==3.0.3",
    "h11==0.14.0",
    "httpcore==1.0.2",
    "httptools==0.6.1",
    "httpx==0.26.0",
    "huggingface-hub==0.20.3",
    "idna==3.6",
    "jinja2==3.1.3",
    "joblib==1.3.2",
    "langchain>=0.3.15",
    "langchain-anthropic==0.3.10",
    "langchain-community>=0.3.15",
    "langchain-core>=0.3.49",
    "langchain-google-genai==2.1.2",
    "langchain-openai>=0.3.1",
    "langchain-text-splitters==0.3.5",
    "langserve==0.3.1",
    "langsmith==0.3.0",
    "llama-index==0.9.48",
    "llama-index-embeddings-openai==0.1.6",
    "lxml==5.1.0",
    "markupsafe==2.1.5",
    "marshmallow==3.20.2",
    "mpmath==1.3.0",
    "multidict==6.0.5",
    "mypy-extensions==1.0.0",
    "nest-asyncio==1.6.0",
    "networkx==3.1",
    "nltk==3.8.1",
    "numpy==1.26.4",
    "openai==1.59.9",
    "packaging==23.2",
    "pandas==2.2.1",
    "pillow==10.2.0",
    "pydantic==2.7.4",
    "pymupdf==1.23.22",
    "pypdf==4.0.1",
    "python-dateutil==2.8.2",
    "python-docx==1.1.2",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "python-pptx==0.6.23",
    "pytz==2024.1",
    "pyyaml==6.0.1",
    "regex==2023.12.25",
    "requests==2.31.0",
    "safetensors==0.4.2",
    "six==1.16.0",
    "sniffio==1.3.0",
    "soupsieve==2.5",
    "sqlalchemy==2.0.27",
    "starlette==0.36.3",
    "supabase>=2.11.0",
    "sympy==1.12",
    "tenacity==8.2.3",
    "tiktoken==0.8.0",
    "tokenizers==0.15.2",
    "torch==2.2.0",
    "tqdm==4.66.2",
    "transformers==4.37.2",
    "typing-inspect==0.9.0",
    "tzdata==2024.1",
    "unstructured==0.12.4",
    "urllib3==2.2.0",
    "uvicorn==0.27.1",
    "watchfiles==0.21.0",
    "websockets==12.0",
    "wrapt==1.16.0",
    "xlsxwriter==3.1.9",
    "yarl==1.9.4",
]
